Scaffold stylesheets, reorganize placeholder app into subdirs
