+++ frontend/config/webpack.config.dev.js
@@ -150,6 +150,7 @@ module.exports = {
               cacheDirectory: true,
             },
           },
+          // Process SCSS
           {
             test: /\.scss$/,
             include: paths.appSrc,
