+++ frontend/src/components/App.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import logo from './logo.svg';
-import './App.scss';
+import logo from '../assets/logo.svg';
 
 // eslint-disable-next-line react/prefer-stateless-function
 class App extends Component {
@@ -10,12 +9,12 @@ class App extends Component {
       <BrowserRouter>
         <Switch>
           <Route exact path="/">
-            <div className="App">
-              <header className="App-header">
-                <img src={logo} className="App-logo" alt="logo" />
-                <h1 className="App-title">Welcome to React</h1>
+            <div className="app">
+              <header className="app-header">
+                <img src={logo} className="app-logo" alt="logo" />
+                <h1 className="app-title">Welcome to React</h1>
               </header>
-              <p className="App-intro">
+              <p className="app-intro">
                 To get started, edit
                 <code>
                   src/App.js
