--- frontend/src/index.scss
@@ -1,8 +1,0 @@
-$least-popular-color-in-the-world: #FF0;
-
-body {
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-  background-color: $least-popular-color-in-the-world;
-}
