+++ frontend/src/styles/base/_all.scss
@@ -1,0 +1,4 @@
+
+@import './base.scss';
+@import './colors.scss';
+@import './typography.scss';
