+++ frontend/src/styles/base/colors.scss
@@ -1,0 +1,5 @@
+/////////////////////////
+//  COLOR DECLARATIONS //
+/////////////////////////
+
+$shadow: rgba(0, 82, 174, 0.15);
