+++ frontend/src/styles/base/typography.scss
@@ -1,0 +1,18 @@
+/////////////////////////
+//        FONTS        //
+/////////////////////////
+
+// todo - font faces
+
+/////////////////////////
+//    TEXT CLASSES     //
+/////////////////////////
+
+// todo - headers, more classes, etc
+
+.text {
+  &-center { text-align: center }
+  &-nowrap { white-space: nowrap }
+}
+
+.underline { text-decoration: underline }
