+++ frontend/src/styles/components/_all.scss
@@ -1,0 +1,1 @@
+@import './app.scss'
