+++ frontend/src/styles/main.scss
@@ -1,0 +1,3 @@
+@import './base/all';
+@import './utils/all';
+@import './components/all';
