+++ frontend/src/styles/utils/_all.scss
@@ -1,0 +1,5 @@
+@import './decorations.scss';
+@import './flex.scss';
+@import './spacing.scss';
+@import './utils.scss';
+@import './breakpoints.scss';
